separate LSTM class
